Rename RequestField namespace
