+++ src/ArgumentResolver/FileSourceRequestResolver.php
@@ -7,8 +7,8 @@ namespace App\ArgumentResolver;
 use App\Entity\AbstractSource;
 use App\Exception\ErrorResponseException;
 use App\Request\FileSourceRequest;
-use App\RequestField\Field\Factory;
-use App\RequestField\Validator\StringFieldValidator;
+use App\RequestParameter\Factory;
+use App\RequestParameter\Validator\StringFieldValidator;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Controller\ValueResolverInterface;
 use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadata;
