+++ src/ArgumentResolver/GitSourceRequestResolver.php
@@ -8,8 +8,8 @@ use App\Entity\AbstractSource;
 use App\Entity\GitSource;
 use App\Exception\ErrorResponseException;
 use App\Request\GitSourceRequest;
-use App\RequestField\Field\Factory;
-use App\RequestField\Validator\StringFieldValidator;
+use App\RequestParameter\Factory;
+use App\RequestParameter\Validator\StringFieldValidator;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Controller\ValueResolverInterface;
 use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadata;
