+++ src/ArgumentResolver/SuiteRequestResolver.php
@@ -10,9 +10,9 @@ use App\Exception\EntityNotFoundException;
 use App\Exception\ErrorResponseException;
 use App\Repository\SourceRepository;
 use App\Request\SuiteRequest;
-use App\RequestField\Field\Factory;
-use App\RequestField\Validator\StringFieldValidator;
-use App\RequestField\Validator\YamlFilenameCollectionFieldValidator;
+use App\RequestParameter\Factory;
+use App\RequestParameter\Validator\StringFieldValidator;
+use App\RequestParameter\Validator\YamlFilenameCollectionFieldValidator;
 use App\Security\EntityAccessChecker;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Controller\ValueResolverInterface;
